# lighty-openapi.yml
openapi: 3.0.0
info:
    title: lighty
    version: 0.0.1
    description: lighty RESTful Stateless API boilerplate documentation
paths:
    /admin/users:
        get:
            summary: Retrieve a list of users
            tags:
                - admin
            security:
                - jwtAuth: []
            responses:
                "200":
                    description: A list of users.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/UserProfile"
                                    count:
                                        type: integer
    /admin/users/{id}:
        get:
            summary: Retrieve a user by ID
            tags:
                - admin
            security:
                - jwtAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                  description: The user ID
            responses:
                "200":
                    description: A user object.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserProfile"
        patch:
            summary: Update a user by ID
            tags:
                - admin
            security:
                - jwtAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                  description: The user ID
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserProperties"
            responses:
                "200":
                    description: The updated user object.
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserProfile"
        delete:
            summary: Delete a user by ID
            tags:
                - admin
            security:
                - jwtAuth: []
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                  description: The user ID
            responses:
                "200":
                    description: deletion acknowledged.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    acknowledged:
                                        type: boolean
                                    deletedCount:
                                        type: integer
    /auth/update-password:
        post:
            summary: update password
            tags:
                - auth
            security:
                - jwtAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                oldPassword:
                                    type: string
                                newPassword:
                                    type: string
            responses:
                "200":
                    description: password updated
                "400":
                    description: incorrect old password
                "401":
                    description: unauthorized
    /:
        get:
            description: welcome to lighty
            responses:
                "200":
                    description: show welcome message
    /ping:
        get:
            description: play ping-pong with server
            responses:
                "200":
                    description: reply with pong.
    /health:
        get:
            description: health check with database connectivity status
            responses:
                "200":
                    description: service is healthy
                "503":
                    description: service is unhealthy
    /favicon.ico:
        get:
            description: get the favicon
            responses:
                "200":
                    description: return favicon.ico
    /users/confirm:
        get:
            summary: confirm user email
            tags:
                - users
            parameters:
                - in: query
                  name: l
                  description: lookup
                  required: true
                  schema:
                      type: string
                - in: query
                  name: v
                  description: verify
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: user verified
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserProfile"
                "400":
                    description: invalid verification
    /users/{id}:
        get:
            summary: get user by id
            tags:
                - users
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: user found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserProfile"
                "404":
                    description: user not found
    /auth/login:
        post:
            summary: authenticate user
            tags:
                - auth
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    format: email
                                    description: the email address of user
                                password:
                                    type: string
                                    format: password
                                    description: the password of user
            responses:
                "200":
                    description: user authenticated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserProfileWithToken"
                "401":
                    description: invalid credentials
    /auth/forgot-password:
        post:
            summary: send reset password email
            tags:
                - auth
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    format: email
                                    description: the email address of user
            responses:
                "200":
                    description: forgot password request received
    /auth/reset-password:
        post:
            summary: reset password
            tags:
                - auth
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                lookup:
                                    type: string
                                verify:
                                    type: string
                                password:
                                    type: string
                                    format: password
                                    description: the password of user
            responses:
                "200":
                    description: password reset
                "400":
                    description: invalid verification
    /users/register:
        post:
            summary: register a new user
            tags:
                - users
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: the name of user
                                email:
                                    type: string
                                    format: email
                                    description: the email address of user
                                password:
                                    type: string
                                    format: password
                                    description: the password of user
            responses:
                "201":
                    description: user created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/UserProfile"
                "400":
                    description: email is already registered
components:
    securitySchemes:
        jwtAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
        UserProperties:
            type: object
            properties:
                name:
                    type: string
                    description: the name of user
                email:
                    type: string
                    format: email
                    description: the email address of user
                role:
                    type: string
                    enum:
                        - admin
                        - user
                    description: the role of user
                confirmed:
                    type: boolean
                    description: the confirmation status of user
        UserProfile:
            type: object
            properties:
                _id:
                    type: string
                    description: the unique identifier of user
                name:
                    type: string
                    description: the name of user
                email:
                    type: string
                    format: email
                    description: the email address of user
                role:
                    type: string
                    enum:
                        - admin
                        - user
                    description: the role of user
                confirmed:
                    type: boolean
                    description: the confirmation status of user
        UserProfileWithToken:
            type: object
            properties:
                _id:
                    type: string
                    description: the unique identifier of user
                name:
                    type: string
                    description: the name of user
                email:
                    type: string
                    format: email
                    description: the email address of user
                role:
                    type: string
                    enum:
                        - admin
                        - user
                    description: the role of user
                confirmed:
                    type: boolean
                    description: the confirmation status of user
                token:
                    type: string
                    description: >-
                        authentication token of user for API access, only available when
                        user is being authenticated
                    x-optional: true
tags: []